pgwire: delete v3Conn
Release note: None
