+++ pkg/sql/pgwire/conn_test.go
@@ -15,7 +15,10 @@
 package pgwire
 
 import (
+	"bytes"
 	"context"
+	"io"
+	"io/ioutil"
 	"net"
 	"strconv"
 	"strings"
@@ -42,6 +45,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/metric"
 	"github.com/cockroachdb/cockroach/pkg/util/mon"
+	"github.com/cockroachdb/cockroach/pkg/util/stop"
 )
 
 // Test the conn struct: check that it marshalls the correct commands to the
@@ -684,4 +688,147 @@ func TestConnClose(t *testing.T) {
 			t.Fatal(err)
 		}
 	})
+}
+
+// TestMaliciousInputs verifies that known malicious inputs sent to
+// a v3Conn don't crash the server.
+func TestMaliciousInputs(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+
+	ctx := context.TODO()
+
+	for _, tc := range [][]byte{
+		// This byte string sends a pgwirebase.ClientMsgClose message type. When
+		// ReadBuffer.readUntypedMsg is called, the 4 bytes is subtracted
+		// from the size, leaving a 0-length ReadBuffer. Following this,
+		// handleClose is called with the empty buffer, which calls
+		// getPrepareType. Previously, getPrepareType would crash on an
+		// empty buffer. This is now fixed.
+		{byte(pgwirebase.ClientMsgClose), 0x00, 0x00, 0x00, 0x04},
+		// This byte string exploited the same bug using a pgwirebase.ClientMsgDescribe
+		// message type.
+		{byte(pgwirebase.ClientMsgDescribe), 0x00, 0x00, 0x00, 0x04},
+		// This would cause ReadBuffer.getInt16 to overflow, resulting in a
+		// negative value being used for an allocation size.
+		{byte(pgwirebase.ClientMsgParse), 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0xff, 0xff},
+	} {
+		t.Run("", func(t *testing.T) {
+			w, r := net.Pipe()
+			defer w.Close()
+			defer r.Close()
+
+			go func() {
+				// This io.Copy will discard all bytes from w until w is closed.
+				// This is needed because sends on the net.Pipe are synchronous, so
+				// the conn will block if we don't read whatever it tries to send.
+				// The reason this works is that ioutil.devNull implements ReadFrom
+				// as an infinite loop, so it will Read continuously until it hits an
+				// error (on w.Close()).
+				_, _ = io.Copy(ioutil.Discard, w)
+			}()
+
+			errChan := make(chan error, 1)
+			go func() {
+				// Write the malicious data.
+				if _, err := w.Write(tc); err != nil {
+					errChan <- err
+					return
+				}
+
+				// Sync and terminate if a panic did not occur to stop the server.
+				// We append a 4-byte trailer to each to signify a zero length message. See
+				// lib/pq.conn.sendSimpleMessage for a similar approach to simple messages.
+				_, _ = w.Write([]byte{byte(pgwirebase.ClientMsgSync), 0x00, 0x00, 0x00, 0x04})
+				_, _ = w.Write([]byte{byte(pgwirebase.ClientMsgTerminate), 0x00, 0x00, 0x00, 0x04})
+				close(errChan)
+			}()
+
+			stopper := stop.NewStopper()
+			defer stopper.Stop(ctx)
+
+			sqlMetrics := sql.MakeMemMetrics("test" /* endpoint */, time.Second /* histogramWindow */)
+			metrics := makeServerMetrics(&sqlMetrics, time.Second /* histogramWindow */)
+
+			conn := newConn(r, sql.SessionArgs{}, &metrics, nil /* execCfg */)
+			// Ignore the error from serveImpl. There might be one when the client
+			// sends malformed input.
+			_ /* err */ = conn.serveImpl(
+				ctx,
+				func() bool { return false }, /* draining */
+				nil,                /* sqlServer */
+				mon.BoundAccount{}, /* reserved */
+				stopper,
+			)
+			if err := <-errChan; err != nil {
+				t.Fatal(err)
+			}
+		})
+	}
+}
+
+// TestReadTimeoutConn asserts that a readTimeoutConn performs reads normally
+// and exits with an appropriate error when exit conditions are satisfied.
+func TestReadTimeoutConnExits(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	// Cannot use net.Pipe because deadlines are not supported.
+	ln, err := net.Listen(util.TestAddr.Network(), util.TestAddr.String())
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer func() {
+		if err := ln.Close(); err != nil {
+			t.Fatal(err)
+		}
+	}()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	expectedRead := []byte("expectedRead")
+
+	// Start a goroutine that performs reads using a readTimeoutConn.
+	errChan := make(chan error)
+	go func() {
+		defer close(errChan)
+		errChan <- func() error {
+			c, err := ln.Accept()
+			if err != nil {
+				return err
+			}
+			defer c.Close()
+
+			readTimeoutConn := newReadTimeoutConn(c, ctx.Err)
+			// Assert that reads are performed normally.
+			readBytes := make([]byte, len(expectedRead))
+			if _, err := readTimeoutConn.Read(readBytes); err != nil {
+				return err
+			}
+			if !bytes.Equal(readBytes, expectedRead) {
+				return errors.Errorf("expected %v got %v", expectedRead, readBytes)
+			}
+
+			// The main goroutine will cancel the context, which should abort
+			// this read with an appropriate error.
+			_, err = readTimeoutConn.Read(make([]byte, 1))
+			return err
+		}()
+	}()
+
+	c, err := net.Dial(ln.Addr().Network(), ln.Addr().String())
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer c.Close()
+
+	if _, err := c.Write(expectedRead); err != nil {
+		t.Fatal(err)
+	}
+
+	select {
+	case err := <-errChan:
+		t.Fatalf("goroutine unexpectedly returned: %v", err)
+	default:
+	}
+	cancel()
+	if err := <-errChan; err != context.Canceled {
+		t.Fatalf("unexpected error: %v", err)
+	}
 }
